renaming wrapper to decomposer
