+++ graphlearn/graphlearn_cipml/graphlearn_cipml.py
@@ -1,5 +1,5 @@
 from graphlearn.graphlearn import GraphLearnSampler
-from graphlearn.graph import core_substitution, graph_clean
+from graphlearn.decompose import core_substitution, graph_clean
 
 from cip_predictor import CipPredictor
