+++ graphlearn/trial_samplers/graphlearn_easy.py
@@ -3,7 +3,7 @@ import itertools
 import random
 import processing
 import estimator_wrapper
-from graphlearn.graph import extract_core_and_interface, core_substitution, graph_clean
+from graphlearn.decompose import extract_core_and_interface, core_substitution, graph_clean
 from graphlearn.feasibility import FeasibilityChecker
 from graphlearn.localsubstitutablegraphgrammar import LocalSubstitutableGraphGrammar
 from multiprocessing import Pool
