+++ graphlearn/utils/interactive.py
@@ -1,5 +1,5 @@
 from graphlearn.processing import PreProcessor
-from graphlearn.graph import Wrapper
+from graphlearn.decompose import Decomposer
 import graphlearn
 
 '''
@@ -7,12 +7,12 @@ first we generate a preprocesser where we can define when the cleaning happens.
 '''
 
 
-class mywrap(Wrapper):
+class mywrap(Decomposer):
     def clean(self):
         return
 
     def real_clean(self):
-        graphlearn.graph.graph_clean(self._base_graph)
+        graphlearn.decompose.graph_clean(self._base_graph)
 
 
 class simple_pp(PreProcessor):
